BackgroundRemover/requirements.txt
rembg==2.0.52
Pillow==10.1.0